tea: Add StartM() that returns final model
Sometimes it is useful to obtain the final model after the
program has finished. Add StartM() (Model, error), so that
it is possible to obtain the final model.

Make sure to keep the original Start() error function, so that
everything is backwards compatible.

The M in StartM was inspired by cobra. Which uses Run and RunE.