Proposal: provide tea.Context during model.Init
This pre-initializes a lipgloss.Renderer on the appropriate output,
which should be used by all Bubbles and components to correctly render
taking the terminal capabilities into account. The renderer is stored
in a tea.Context, which gets passed to each model on initialization.
This could also hold other properties like the initial dimensions of
the output or settings like alt-screen or bracketed paste being
supported.

This is a breaking API change as it changes tea.Model's interface.