feat: add color profiles and detect the terminal profile on program run
This defines `tea.Profile` and its consts. A color profile can downgrade
`color.Color` based on terminal colors support. When the program starts,
it will detect the color profile from the output and the program
environment variables before sending the profile to the program model.

This also adds a `startup` example that demonstrates messages a program
receive during startup.