refactor: move input tracing to input parser and simplify tracing
This commit wraps the input parser in a struct and moves the tracing
logic to the program and its writer and input driver.

It also changes the behavior of the tracer to only enable tracing for
local environment variables i.e. from `os.Getenv` and not from the
program's environment. In other words, it will ignore the program's
environment variables when it comes to tracing.