fix: renderer: keep a separate count of lines rendered in the alt screen
We need to keep a separate count of lines rendered in the alt screen and
inline mode. This is because when we enter alt screen mode, the cursor
position is saved and the alt screen is cleared. Now when we exit the
alt screen mode, the cursor position is restored so we don't need to
move the cursor to the beginning of the section that we rendered.

To fix this, we keep a separate count of lines rendered in the alt screen
mode and inline mode.

Related: https://github.com/charmbracelet/bubbletea/issues/1013
Fixes: https://github.com/charmbracelet/bubbletea/pull/1241
Fixes: https://github.com/charmbracelet/bubbletea/issues/1248