feat: add environ type and EnvMsg message (#1287)
Send the environment variables used by the program to the program when it
starts. This is done by sending an `EnvMsg` message to the program.
Programs that need to access the environment variables can use this
message type to get the values of the environment variables.

This is useful when programs targeting remote sessions need to access
environment variables that are set by the user or the shell.