fix: use CancelIoEx to cancel Windows conInputReader across threads
The Windows conInputReader uses `windows.CancelIo` to close the read
loop's input handle. Windows `CancelIo` docs indicate that call is
only appropriate for I/O operations on the same thread as the `CancelIo`
call. Since the read loop is in its own goroutine, it's likely to be
on another thread from the bubbletea Program.

This change adds a call to Windows `CancelIoEx`, which will cancel I/O
operations across threads.