fix: don't send keyboard enhancements message on startup
This was a hack to detect if the terminal supports keyboard
enhancements. It hardcoded a timeout of 100ms to wait for the keyboard
enhancements to be read from the terminal before sending an empty
message. This is wrong because it assumes a fixed timeout for all
terminals. While this is fine for now, we would never know if this
timeout works for all terminals or if it is too short for something like
an SSH session.

Thus, we remove this hack and let the program handle message detection.
A program can simply have a custom timeout message that gets fired after
a certain time after sending all the terminal query commands.