fix(renderer): properly reset cursor position to start of line
This change uses carriage return (`\r`) instead of moving the cursor
backward by the terminal width at the end of rendering in inline mode.
With this, we don't need to know the terminal width to reset the cursor
position, which avoids issues when the renderer doesn't yet know the
terminal width (e.g. the first render where the size message hasn't been
sent to the renderer yet).

Fixes: https://github.com/charmbracelet/bubbletea/issues/1471